Fix (merge) migrations after rebase
